refactor "item shipped"
