+++ src/main/java/org/folio/innreach/domain/service/CirculationService.java
@@ -2,6 +2,7 @@ package org.folio.innreach.domain.service;
 
 import org.folio.innreach.dto.CirculationRequestDTO;
 import org.folio.innreach.dto.InnReachResponseDTO;
+import org.folio.innreach.dto.ItemShippedDTO;
 import org.folio.innreach.dto.PatronHoldDTO;
 import org.folio.innreach.dto.TransferRequestDTO;
 
@@ -11,5 +12,8 @@ public interface CirculationService {
 
   InnReachResponseDTO initiatePatronHold(String trackingId, String centralCode, PatronHoldDTO patronHold);
 
+  InnReachResponseDTO trackShippedItem(String trackingId, String centralCode, ItemShippedDTO itemShipped);
+
   InnReachResponseDTO transferItem(String trackingId, String centralCode, TransferRequestDTO transferRequest);
+
 }
