+++ src/main/resources/swagger.api/circulation.yaml
@@ -182,7 +182,36 @@ paths:
               $ref: "#/components/schemas/patronHoldDTO"
       responses:
         '200':
-          description: Transfer request complete
+          description: Patron hold request complete
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/innReachResponseDTO"
+        '400':
+          description: An error occurred during processing the request
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/innReachResponseDTO"
+        '500':
+          $ref: "api-common.yaml#/components/responses/trait_response_500"
+      parameters:
+        - $ref: '#/components/parameters/trackingId'
+        - $ref: '#/components/parameters/centralCode'
+  /d2ir/circ/itemshipped/{trackingId}/{centralCode}:
+    put:
+      operationId: itemShipped
+      description: Process shipped item
+      tags:
+        - inn-reach-circulation
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: "#/components/schemas/itemShippedDTO"
+      responses:
+        '200':
+          description: Item shipped request complete
           content:
             application/json:
               schema:
@@ -293,6 +322,8 @@ components:
       $ref: schemas/d2ir/circulationRequestDTO.json
     patronHoldDTO:
       $ref: schemas/d2ir/patronHoldDTO.json
+    itemShippedDTO:
+      $ref: schemas/d2ir/itemShippedDTO.json
     transferRequestDTO:
       $ref: schemas/d2ir/transferRequestDTO.json
     innReachTransactionsDTO:
