+++ src/test/java/org/folio/innreach/controller/d2ir/PatronHoldCirculationApiTest.java
@@ -14,7 +14,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import static org.folio.innreach.domain.CirculationOperation.CANCEL_PATRON_HOLD;
-import static org.folio.innreach.domain.CirculationOperation.PATRON_HOLD;
 import static org.folio.innreach.fixture.CirculationFixture.createTransactionHoldDTO;
 import static org.folio.innreach.util.UUIDHelper.toStringWithoutHyphens;
 
@@ -66,6 +65,7 @@ class PatronHoldCirculationApiTest extends BaseApiControllerTest {
   private static final String INSTANCE_TYPE_NAME_URLENCODED = "INN-Reach%20temporary%20record";
   private static final String INSTANCE_CONTRIBUTOR_NAME_URLENCODED = "INN-Reach%20author";
 
+  private static final String PATRON_HOLD_OPERATION = "patronhold";
   private static final String CIRCULATION_ENDPOINT = "/inn-reach/d2ir/circ/{circulationOperationName}/{trackingId}/{centralCode}";
 
   private static final String PRE_POPULATED_TRACKING_ID = "tracking1";
@@ -116,7 +116,7 @@ class PatronHoldCirculationApiTest extends BaseApiControllerTest {
     stubGet(format(QUERY_REQUEST_BY_ITEM_ID_URL_TEMPLATE, PRE_POPULATED_ITEM_ID), "circulation/empty-requests-response.json");
     stubPost(format(MOVE_CIRCULATION_REQUEST_URL_TEMPLATE, PRE_POPULATED_REQUEST_ID), "circulation/updated-request-response.json");
 
-    mockMvc.perform(post(CIRCULATION_ENDPOINT, PATRON_HOLD.getOperationName(), PRE_POPULATED_TRACKING_ID, PRE_POPULATED_CENTRAL_CODE)
+    mockMvc.perform(post(CIRCULATION_ENDPOINT, PATRON_HOLD_OPERATION, PRE_POPULATED_TRACKING_ID, PRE_POPULATED_CENTRAL_CODE)
         .content(jsonHelper.toJson(transactionHoldDTO))
         .contentType(MediaType.APPLICATION_JSON)
         .headers(getOkapiHeaders()))
@@ -156,7 +156,7 @@ class PatronHoldCirculationApiTest extends BaseApiControllerTest {
     stubGet(format(QUERY_INVENTORY_ITEM_BY_HRID_URL_TEMPLATE, ITEM_HRID), "inventory/query-items-response.json");
     stubGet(format(QUERY_REQUEST_BY_ITEM_ID_URL_TEMPLATE, PRE_POPULATED_ITEM_ID), "circulation/empty-requests-response.json");
 
-    mockMvc.perform(post(CIRCULATION_ENDPOINT, PATRON_HOLD.getOperationName(), "newtrackingid", PRE_POPULATED_CENTRAL_CODE)
+    mockMvc.perform(post(CIRCULATION_ENDPOINT, PATRON_HOLD_OPERATION, "newtrackingid", PRE_POPULATED_CENTRAL_CODE)
         .content(jsonHelper.toJson(transactionHoldDTO))
         .contentType(MediaType.APPLICATION_JSON)
         .headers(getOkapiHeaders()))
